# docs/api/dgraph-openapi.yaml
openapi: 3.0.3
info:
  title: DGraph Service API
  description: |
    Off-chain and social data. All social interactions (posts, feed, interactions),
    prediction markets (list, bets, payouts), governance variables, subscription status,
    location preferences and spots, ad campaigns/impressions/analytics, and safety
    run through this service (read/write from Dgraph or in-memory fallback).
  version: 1.0.0
servers:
  - url: http://localhost:3003
    description: Local default
  - url: https://gql.dlux.io
    description: Test/production
...
paths:
...
  /social/posts:
    get:
      tags: [social]
      summary: Get feed (social posts)
      parameters:
        - name: author
          in: query
          schema: { type: string }
        - name: dappId
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        '200':
          description: List of posts / feed
    post:
      tags: [social]
      summary: Create post (signed)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [author, content, signature]
              properties:
                author: { type: string }
                content: { type: string }
                signature: { type: string }
                dappId: { type: string }
                contentType: { type: string, default: text }
      responses:
        '201':
          description: Post created
        '400':
          description: Missing author, content, or signature
        '401':
          description: Invalid signature
  /social/posts/{id}:
    get:
      tags: [social]
      summary: Get post by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Post
        '404':
          description: Post not found
  /social/interactions:
    post:
      tags: [social]
      summary: Create interaction (like, repost, reply, etc.)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type: { type: string, enum: [like, dislike, repost, quote, reply] }
                author: { type: string }
                targetId: { type: string }
                targetType: { type: string }
                content: { type: string }
                signature: { type: string }
      responses:
        '201':
          description: Interaction created
        '400':
          description: Bad request
...
  /markets/high-payout:
    get:
      tags: [markets]
      summary: List high-payout (open) prediction markets
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 10 }
      responses:
        '200':
          description: markets, limit
  /markets/dapp/{dappId}:
    get:
      tags: [markets]
      summary: Active markets for a dApp
      parameters:
        - name: dappId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: markets array
  /markets/payouts/{owner}:
    get:
      tags: [markets]
      summary: Payout balance for owner
      parameters:
        - name: owner
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: total (and per-market if applicable)
  /markets:
    post:
      tags: [markets]
      summary: Create prediction market (returns tx id; on-chain creation separate)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                dappId: { type: string }
                safetyMetric: { type: string }
      responses:
        '200':
          description: transactionId
  /markets/{marketId}/bets:
    post:
      tags: [markets]
      summary: Place bet (returns tx id; on-chain bet separate)
      parameters:
        - name: marketId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                bettor: { type: string }
                side: { type: string, enum: [safe, unsafe] }
                amount: { type: number }
      responses:
        '200':
          description: transactionId
...